refactor, move getAdmins fn to auth
